+++ CHANGES.rst
@@ -15,6 +15,7 @@ Unreleased
   in :attr:`Feed.last_exception`. (:issue:`68`)
 * Show details about feed update errors in the web application. (:issue:`68`)
 * Expose the :attr:`~Feed.added` and :attr:`~Feed.last_updated` Feed attributes.
+* Expose the :attr:`~Entry.last_updated` Entry attribute.
 
 
 Version 1.2